+++ src/Burrow.Tests/RabbitTunnelTests/MethodSubscribe.cs
@@ -22,7 +22,7 @@ namespace Burrow.Tests.RabbitTunnelTests
             tunnel.Subscribe<Customer>("subscriptionName", x => { });
 
             // Assert
-            newChannel.Received().BasicQos(0, Global.PreFetchSize, false);
+            newChannel.Received().BasicQos(0, (ushort)Global.PreFetchSize, false);
             newChannel.Received().BasicConsume("Queue", false, Arg.Is<string>(x => x.StartsWith("subscriptionName-")), Arg.Any<IBasicConsumer>());
         }
 
@@ -39,7 +39,7 @@ namespace Burrow.Tests.RabbitTunnelTests
             var subs = tunnel.Subscribe<Customer>("subscriptionName", (x, y) => { });
 
             // Assert
-            newChannel.Received().BasicQos(0, Global.PreFetchSize, false);
+            newChannel.Received().BasicQos(0, (ushort)Global.PreFetchSize, false);
             newChannel.Received().BasicConsume("Queue", false, Arg.Is<string>(x => x.StartsWith("subscriptionName-")), Arg.Any<IBasicConsumer>());
             Assert.IsInstanceOfType(subs, typeof(Subscription));
         }
